Initial BaseKernel.
