+++ spyder_kernels/base/kernel.py
@@ -1,0 +1,372 @@
+# -*- coding: utf-8 -*-
+# -----------------------------------------------------------------------------
+# Copyright (c) 2018- Spyder Kernels Contributors
+#
+# Licensed under the terms of the MIT License
+# (see spyder_kernels/__init__.py for details)
+# -----------------------------------------------------------------------------
+
+"""
+Base Spyder kernel
+"""
+
+# Standard library imports
+import os
+import sys
+
+# Excluded variables from the Variable Explorer (i.e. they are not
+# shown at all there)
+EXCLUDED_NAMES = ['In', 'Out', 'exit', 'get_ipython', 'quit']
+
+# To be able to get and set variables between Python 2 and 3
+PICKLE_PROTOCOL = 2
+
+PY2 = sys.version[0] == '2'
+
+
+class BaseKernelMixIn:
+    """Base Spyder kernel with methods to interact with Spyder."""
+
+    namespace_view_settings = {}
+    _pdb_obj = None
+    _pdb_step = None
+    _do_publish_pdb_state = True
+
+    # -- Public API ----------------------------------------------------
+    # --- For the Variable Explorer
+    def get_namespace_view(self):
+        """
+        Return the namespace view
+
+        This is a dictionary with the following structure
+
+        {'a': {'color': '#800000', 'size': 1, 'type': 'str', 'view': '1'}}
+
+        Here:
+        * 'a' is the variable name
+        * 'color' is the color used to show it
+        * 'size' and 'type' are self-evident
+        * and'view' is its value or the text shown in the last column
+        """
+        from spyder_kernels.utils.nsview import make_remote_view
+
+        settings = self.namespace_view_settings
+        if settings:
+            ns = self._get_current_namespace()
+            view = repr(make_remote_view(ns, settings, EXCLUDED_NAMES))
+            return view
+        else:
+            return repr(None)
+
+    def get_var_properties(self):
+        """
+        Get some properties of the variables in the current
+        namespace
+        """
+        from spyder_kernels.utils.nsview import get_remote_data
+
+        settings = self.namespace_view_settings
+        if settings:
+            ns = self._get_current_namespace()
+            data = get_remote_data(ns, settings, mode='editable',
+                                   more_excluded_names=EXCLUDED_NAMES)
+
+            properties = {}
+            for name, value in list(data.items()):
+                properties[name] = {
+                    'is_list':  isinstance(value, (tuple, list)),
+                    'is_dict':  isinstance(value, dict),
+                    'is_set': isinstance(value, set),
+                    'len': self._get_len(value),
+                    'is_array': self._is_array(value),
+                    'is_image': self._is_image(value),
+                    'is_data_frame': self._is_data_frame(value),
+                    'is_series': self._is_series(value),
+                    'array_shape': self._get_array_shape(value),
+                    'array_ndim': self._get_array_ndim(value)
+                }
+
+            return repr(properties)
+        else:
+            return repr(None)
+
+    def send_spyder_msg(self, spyder_msg_type, content=None, data=None):
+        """
+        Publish custom messages to the Spyder frontend.
+
+        Parameters
+        ----------
+
+        spyder_msg_type: str
+            The spyder message type
+        content: dict
+            The (JSONable) content of the message
+        data: any
+            Any object that is serializable by cloudpickle (should be most
+            things). Will arrive as cloudpickled bytes in `.buffers[0]`.
+        """
+        import cloudpickle
+
+        if content is None:
+            content = {}
+        content['spyder_msg_type'] = spyder_msg_type
+        msg = self.session.send(
+            self.iopub_socket,
+            'spyder_msg',
+            content=content,
+            buffers=[cloudpickle.dumps(data, protocol=PICKLE_PROTOCOL)],
+            parent=self._parent_header,
+        )
+        self.log.debug(msg)
+
+    def get_value(self, name):
+        """Get the value of a variable"""
+        ns = self._get_current_namespace()
+        value = ns[name]
+        try:
+            self.send_spyder_msg('data', data=value)
+        except Exception:
+            # * There is no need to inform users about
+            #   these errors.
+            # * value = None makes Spyder to ignore
+            #   petitions to display a value
+            self.send_spyder_msg('data', data=None)
+        self._do_publish_pdb_state = False
+
+    def set_value(self, name, value, PY2_frontend):
+        """Set the value of a variable"""
+        import cloudpickle
+        ns = self._get_reference_namespace(name)
+
+        # We send serialized values in a list of one element
+        # from Spyder to the kernel, to be able to send them
+        # at all in Python 2
+        svalue = value[0]
+
+        # We need to convert svalue to bytes if the frontend
+        # runs in Python 2 and the kernel runs in Python 3
+        if PY2_frontend and not PY2:
+            svalue = bytes(svalue, 'latin-1')
+
+        # Deserialize and set value in namespace
+        dvalue = cloudpickle.loads(svalue)
+        ns[name] = dvalue
+        self.log.debug(ns)
+
+    def remove_value(self, name):
+        """Remove a variable"""
+        ns = self._get_reference_namespace(name)
+        ns.pop(name)
+
+    def copy_value(self, orig_name, new_name):
+        """Copy a variable"""
+        ns = self._get_reference_namespace(orig_name)
+        ns[new_name] = ns[orig_name]
+
+    def load_data(self, filename, ext):
+        """Load data from filename"""
+        from spyder_kernels.utils.iofuncs import iofunctions
+        from spyder_kernels.utils.misc import fix_reference_name
+
+        glbs = self._mglobals()
+
+        load_func = iofunctions.load_funcs[ext]
+        data, error_message = load_func(filename)
+
+        if error_message:
+            return error_message
+
+        for key in list(data.keys()):
+            new_key = fix_reference_name(key, blacklist=list(glbs.keys()))
+            if new_key != key:
+                data[new_key] = data.pop(key)
+
+        try:
+            glbs.update(data)
+        except Exception as error:
+            return str(error)
+
+        return None
+
+    def save_namespace(self, filename):
+        """Save namespace into filename"""
+        from spyder_kernels.utils.nsview import get_remote_data
+        from spyder_kernels.utils.iofuncs import iofunctions
+
+        ns = self._get_current_namespace()
+        settings = self.namespace_view_settings
+        data = get_remote_data(ns, settings, mode='picklable',
+                               more_excluded_names=EXCLUDED_NAMES).copy()
+        return iofunctions.save(data, filename)
+
+    # --- For the Help plugin
+    def is_defined(self, obj, force_import=False):
+        """Return True if object is defined in current namespace"""
+        from spyder_kernels.utils.dochelpers import isdefined
+
+        ns = self._get_current_namespace(with_magics=True)
+        return isdefined(obj, force_import=force_import, namespace=ns)
+
+    def get_doc(self, objtxt):
+        """Get object documentation dictionary"""
+        try:
+            import matplotlib
+            matplotlib.rcParams['docstring.hardcopy'] = True
+        except Exception:
+            pass
+        from spyder_kernels.utils.dochelpers import getdoc
+
+        obj, valid = self._eval(objtxt)
+        if valid:
+            return getdoc(obj)
+
+    def get_source(self, objtxt):
+        """Get object source"""
+        from spyder_kernels.utils.dochelpers import getsource
+
+        obj, valid = self._eval(objtxt)
+        if valid:
+            return getsource(obj)
+
+    # --- Additional methods
+    def set_cwd(self, dirname):
+        """Set current working directory."""
+        os.chdir(dirname)
+
+    def get_cwd(self):
+        """Get current working directory."""
+        return os.getcwd()
+
+    def get_syspath(self):
+        """Return sys.path contents."""
+        return sys.path[:]
+
+    def get_env(self):
+        """Get environment variables."""
+        return os.environ.copy()
+
+    # -- Private API ---------------------------------------------------
+    # --- For the Variable Explorer
+    def _get_current_namespace(self, with_magics=False):
+        """
+        Return current namespace
+
+        This is globals() if not debugging, or a dictionary containing
+        both locals() and globals() for current frame when debugging
+        """
+        ns = {}
+        glbs = self._mglobals()
+
+        if self._pdb_frame is None:
+            ns.update(glbs)
+        else:
+            ns.update(glbs)
+            ns.update(self._pdb_locals)
+
+        # Add magics to ns so we can show help about them on the Help
+        # plugin
+        if with_magics:
+            line_magics = self.shell.magics_manager.magics['line']
+            cell_magics = self.shell.magics_manager.magics['cell']
+            ns.update(line_magics)
+            ns.update(cell_magics)
+
+        return ns
+
+    def _get_reference_namespace(self, name):
+        """
+        Return namespace where reference name is defined
+
+        It returns the globals() if reference has not yet been defined
+        """
+        glbs = self._mglobals()
+        if self._pdb_frame is None:
+            return glbs
+        else:
+            lcls = self._pdb_locals
+            if name in lcls:
+                return lcls
+            else:
+                return glbs
+
+    def _mglobals(self):
+        """Return current globals -- handles Pdb frames"""
+        if self._pdb_frame is not None:
+            return self._pdb_frame.f_globals
+        else:
+            return self.shell.user_ns
+
+    def _get_len(self, var):
+        """Return sequence length"""
+        try:
+            return len(var)
+        except Exception:
+            return None
+
+    def _is_array(self, var):
+        """Return True if variable is a NumPy array"""
+        try:
+            import numpy
+            return isinstance(var, numpy.ndarray)
+        except Exception:
+            return False
+
+    def _is_image(self, var):
+        """Return True if variable is a PIL.Image image"""
+        try:
+            from PIL import Image
+            return isinstance(var, Image.Image)
+        except Exception:
+            return False
+
+    def _is_data_frame(self, var):
+        """Return True if variable is a DataFrame"""
+        try:
+            from pandas import DataFrame
+            return isinstance(var, DataFrame)
+        except Exception:
+            return False
+
+    def _is_series(self, var):
+        """Return True if variable is a Series"""
+        try:
+            from pandas import Series
+            return isinstance(var, Series)
+        except Exception:
+            return False
+
+    def _get_array_shape(self, var):
+        """Return array's shape"""
+        try:
+            if self._is_array(var):
+                return var.shape
+            else:
+                return None
+        except Exception:
+            return None
+
+    def _get_array_ndim(self, var):
+        """Return array's ndim"""
+        try:
+            if self._is_array(var):
+                return var.ndim
+            else:
+                return None
+        except Exception:
+            return None
+
+    # --- For the Help plugin
+    def _eval(self, text):
+        """
+        Evaluate text and return (obj, valid)
+        where *obj* is the object represented by *text*
+        and *valid* is True if object evaluation did not raise any exception
+        """
+        from spyder_kernels.py3compat import is_text_string
+
+        assert is_text_string(text)
+        ns = self._get_current_namespace(with_magics=True)
+        try:
+            return eval(text, ns), True
+        except Exception:
+            return None, False
